requirements.txt
streamlit
streamlit-option-menu
requests
pillow
google-generativeai
pyyaml